pyproject.toml
[project]
requires-python = "==3.13.*"
name = "songlinker"
description = "Links songs"
version = "3.0.0"
authors = [{ name = "BlindfoldedSurgery", email = "team@blindfolded.surgery" }]
classifiers = [
    "Private :: Do Not Upload",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    "bs-config [dotenv] ==1.1.1",
    "bs-nats-updater ==2.0.0",
    "click >=8.1.3, <9.0.0",
    "httpx >=0.28.1, <0.29.0",
    "opentelemetry-api ==1.34.*",
    "opentelemetry-sdk ==1.34.*",
    "opentelemetry-exporter-otlp-proto-grpc ==1.34.*",
    "opentelemetry-instrumentation-asyncio",
    "opentelemetry-instrumentation-httpx",
    "opentelemetry-instrumentation-logging",
    "pydantic ==2.11.*",
    "python-telegram-bot ==22.1",
    "sentry-sdk >=2.0.0, <3.0.0",
    "uvloop ==0.21.*",
]

[dependency-groups]
dev = [
    "pyright ==1.1.402",
    "pytest >=8.0.0, <9.0.0",
    "pytest-asyncio==1.0.0",
    "pytest-mock >=3.11.1, <4.0.0",
    "pytest-recording >=0.13.0, <0.14.0",
    "ruff ==0.12.1",
    "vcrpy >=7.0.0, <7.1.0",
]

[[tool.uv.index]]
name = "internal"
url = "https://pypi.bjoernpetersen.net"
explicit = true

[tool.uv.sources]
bs-config = { index = "internal" }
bs-nats-updater = { index = "internal" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = [
    "integration",
]

[tool.pyright]
strict = [ "src/songlinker/**/*" ]

[tool.ruff.lint]
select = [
    "E4",
    "E7",
    "E9",
    "F",
    "FURB",
    "G",
    "I001",
    "LOG",
    "PTH",
    "Q",
    "UP",
]
ignore = [
    # redundant-open-modes
    "UP015",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["songlinker"]
